requirements.txt
python-dotenv==1.0.1
pandas==2.2.3
numpy==2.1.3
requests==2.32.3
pyarrow==17.0.0
scikit-learn==1.5.2
scipy==1.13.1
matplotlib==3.9.2
seaborn==0.13.2
streamlit==1.39.0
pandera==0.20.3
PyYAML~=6.0.2
toml~=0.10.2